+++ src/backend/nsl/nsl_interp.h
@@ -49,7 +49,7 @@ extern const char* nsl_interp_evaluate_name[];
 
 /* calculates integration of n points of xy-data. result in y */
 /* TODO: put in nsl_int.h */
-int nsl_interp_integral(double *x, double *y, unsigned int n);
+int nsl_interp_integral(double *x, double *y, size_t n);
 
 /* calculates rational interpolation of n points of xy-data at xn using Burlisch-Stoer method. result in v (error dv) */
 int nsl_interp_ratint(double *x, double *y, int n, double xn, double *v, double *dv);
